# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "MileApp Screening Test",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/package/": {
            "get": {
                "description": "Get List Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Get List Package",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Create Package",
                "parameters": [
                    {
                        "description": "Create Package",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.TransactionCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/package/{id}": {
            "get": {
                "description": "Get Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Get Package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Update Package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Package",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.TransactionUpdateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Delete Package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch Package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Patch Package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patch Package",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.TransactionPatchParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "params.ConnoteCreateParam": {
            "type": "object",
            "properties": {
                "actual_weight": {
                    "type": "integer"
                },
                "chargeable_weight": {
                    "type": "integer"
                },
                "connote_amount": {
                    "type": "integer"
                },
                "connote_booking_code": {
                    "type": "string"
                },
                "connote_count": {
                    "type": "string"
                },
                "connote_number": {
                    "type": "integer"
                },
                "connote_order": {
                    "type": "integer"
                },
                "connote_service": {
                    "type": "string"
                },
                "connote_service_price": {
                    "type": "integer"
                },
                "connote_sla_day": {
                    "type": "string"
                },
                "connote_state": {
                    "type": "string"
                },
                "connote_state_id": {
                    "type": "integer"
                },
                "connote_surcharge_amount": {
                    "type": "string"
                },
                "connote_total_package": {
                    "type": "string"
                },
                "history": {
                    "type": "array",
                    "items": {}
                },
                "id_source_tarif": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string"
                },
                "location_type": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "pod": {
                    "type": "string"
                },
                "source_tarif_db": {
                    "type": "string"
                },
                "surcharge_amount": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "string"
                },
                "volume_weight": {
                    "type": "integer"
                },
                "zone_code_from": {
                    "type": "string"
                },
                "zone_code_to": {
                    "type": "string"
                }
            }
        },
...
        "params.KoliDataCreateParam": {
            "type": "object",
            "properties": {
                "awb_url": {
                    "type": "string"
                },
                "connote_id": {
                    "type": "string"
                },
                "koli_chargeable_weight": {
                    "type": "integer"
                },
                "koli_code": {
                    "type": "string"
                },
                "koli_custom_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "koli_description": {
                    "type": "string"
                },
                "koli_formula_id": {
                    "type": "integer"
                },
                "koli_height": {
                    "type": "integer"
                },
                "koli_length": {
                    "type": "integer"
                },
                "koli_surcharge": {
                    "type": "array",
                    "items": {}
                },
                "koli_volume": {
                    "type": "integer"
                },
                "koli_weight": {
                    "type": "integer"
                },
                "koli_width": {
                    "type": "integer"
                }
            }
        },
        "params.KoliDataUpdateParam": {
            "type": "object",
            "properties": {
                "awb_url": {
                    "type": "string"
                },
                "connote_id": {
                    "type": "string"
                },
                "koli_chargeable_weight": {
                    "type": "integer"
                },
                "koli_code": {
                    "type": "string"
                },
                "koli_custom_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "koli_description": {
                    "type": "string"
                },
                "koli_formula_id": {
                    "type": "integer"
                },
                "koli_height": {
                    "type": "integer"
                },
                "koli_id": {
                    "type": "string"
                },
                "koli_length": {
                    "type": "integer"
                },
                "koli_surcharge": {
                    "type": "array",
                    "items": {}
                },
                "koli_volume": {
                    "type": "integer"
                },
                "koli_weight": {
                    "type": "integer"
                },
                "koli_width": {
                    "type": "integer"
                }
            }
        },
        "params.TransactionCreateParam": {
            "type": "object",
            "required": [
                "current_location",
                "customer_code",
                "customer_name",
                "location_id",
                "organization_id",
                "transaction_amount",
                "transaction_cash_amount",
                "transaction_cash_change",
                "transaction_discount",
                "transaction_payment_type",
                "transaction_payment_type_name",
                "transaction_state"
            ],
            "properties": {
                "connote": {
                    "$ref": "#/definitions/params.ConnoteCreateParam"
                },
                "connote_id": {
                    "type": "string"
                },
                "current_location": {
                    "$ref": "#/definitions/params.TransactionCurrentLocationCreateParam"
                },
                "custom_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_attribute": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_code": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "destination_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "koli_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/params.KoliDataCreateParam"
                    }
                },
                "location_id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "origin_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "transaction_additional_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "transaction_amount": {
                    "type": "string"
                },
                "transaction_cash_amount": {
                    "type": "integer"
                },
                "transaction_cash_change": {
                    "type": "integer"
                },
                "transaction_discount": {
                    "type": "string"
                },
                "transaction_payment_type": {
                    "type": "string"
                },
                "transaction_payment_type_name": {
                    "type": "string"
                },
                "transaction_state": {
                    "type": "string"
                }
            }
        },
        "params.TransactionCurrentLocationCreateParam": {
            "type": "object",
            "required": [
                "code",
                "name",
                "type"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "params.TransactionCurrentLocationPatchParam": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "params.TransactionCustomerCreateParam": {
            "type": "object",
            "properties": {
                "customer_address": {
                    "type": "string"
                },
                "customer_address_detail": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "customer_phone": {
                    "type": "string"
                },
                "customer_zip_code": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "zone_code": {
                    "type": "string"
                }
            }
        },
        "params.TransactionPatchParam": {
            "type": "object",
            "properties": {
                "connote": {
                    "$ref": "#/definitions/params.ConnoteUpdateParam"
                },
                "connote_id": {
                    "type": "string"
                },
                "current_location": {
                    "$ref": "#/definitions/params.TransactionCurrentLocationPatchParam"
                },
                "custom_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_attribute": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_code": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "destination_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "koli_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/params.KoliDataUpdateParam"
                    }
                },
                "location_id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "origin_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "transaction_additional_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "transaction_amount": {
                    "type": "string"
                },
                "transaction_cash_amount": {
                    "type": "integer"
                },
                "transaction_cash_change": {
                    "type": "integer"
                },
                "transaction_discount": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_payment_type": {
                    "type": "string"
                },
                "transaction_payment_type_name": {
                    "type": "string"
                },
                "transaction_state": {
                    "type": "string"
                }
            }
        },
        "params.TransactionUpdateParam": {
            "type": "object",
            "required": [
                "current_location",
                "customer_code",
                "customer_name",
                "location_id",
                "organization_id",
                "transaction_amount",
                "transaction_cash_amount",
                "transaction_cash_change",
                "transaction_discount",
                "transaction_payment_type",
                "transaction_payment_type_name",
                "transaction_state"
            ],
            "properties": {
                "connote": {
                    "$ref": "#/definitions/params.ConnoteUpdateParam"
                },
                "connote_id": {
                    "type": "string"
                },
                "current_location": {
                    "$ref": "#/definitions/params.TransactionCurrentLocationCreateParam"
                },
                "custom_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_attribute": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customer_code": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "destination_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "koli_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/params.KoliDataUpdateParam"
                    }
                },
                "location_id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "origin_data": {
                    "$ref": "#/definitions/params.TransactionCustomerCreateParam"
                },
                "transaction_additional_field": {
                    "type": "object",
                    "additionalProperties": true
                },
                "transaction_amount": {
                    "type": "string"
                },
                "transaction_cash_amount": {
                    "type": "integer"
                },
                "transaction_cash_change": {
                    "type": "integer"
                },
                "transaction_discount": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "transaction_payment_type": {
                    "type": "string"
                },
                "transaction_payment_type_name": {
                    "type": "string"
                },
                "transaction_state": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}